fix(ci): rename queue package to job_queue
Renamed agent_brain_server/queue to agent_brain_server/job_queue to
avoid conflict with Python's built-in queue module. This conflict caused
Poetry import errors during CI installation.
